requirements.txt
alembic==1.8.0
anyio==3.6.1
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
coverage==6.4.1
dnspython==2.2.1
email-validator==1.2.1
execnet==1.9.0
fastapi==0.78.0
FastAPI-SQLAlchemy==0.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dotenv==0.20.0
rfc3986==1.5.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
